Rewrite sound cache management
Instead of swapping actual sound sources around, we now replace the
sound buffers in the sound sources from the ones in the cache.

This is a much better approach because if a sound source is modified
beyond its own class defaults (loop/no loop etc), these changes stay
with the source and don't need to be redone after a change in buffer.

Sadly this means adding a new field to sound sources to ease their
management. On the upside, we are not flooded by missing sound file
messages every time the same (missing) sound file is selected, and the
sound source treatment is unified across all kinds of sound sources,
with a common cache.