No source check in ambient/player sounds init
This was initially done as a precaution against trying to set the AL
source behavior for a non-existing sources, but now the source is
always created, even without buffers.

And in fact, many sounds can be created without an initial buffer,
resulting in their properties not being correctly set up (this can be
seen by e.g. setting a mono ambient sound and discovering it doesn't
move with the player).

Remove the source check, allowing ambient and player sounds to always
have the correct initial properties.